funding+rpcserver+log: move funding manager to own package
To not clutter the main package namespace with too much functionality,
we move the funding manager into its own package.
